DomGenie_BackEnd_V2-main-UPDATED/DomGenie_BackEnd_V2-main/requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.97.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.1
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pytest==7.3.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
asyncmy
requests
Flask
google-api-python-client==2.98.0
google-auth-httplib2==0.1.0
google-auth
google-auth-oauthlib
PyPDF2
python-magic
pdfkit
wkhtmltopdf
UUID
reportlab
python-docx
fpdf
boto3
pillow 
python-pptx
docx2pdf
gunicorn
jinja2
